students/y2333/laboratory_works/agapova_ekaterina/laboratiry_work_1/requirements.txt
Django==3.0.7
djangorestframework
djangorestframework-jwt
django-webpack-loader==0.6.0